Make test.sh a bit nicer
Update test.sh to naively detect whether or not you've ran bootstrap.sh.

Also update it to source the environment for you.

+++ test.sh
@@ -1,5 +1,17 @@
 #!/usr/bin/env bash
 
+if [ ! -d "env" ]; then
+    echo "Could not find Python env. Have you ran bootstrap.sh?"
+    exit 1
+fi
+
+if [ ! -d "vendor/" ]; then
+    echo "Could not find Go vendor directory. Have you ran bootstrap.sh?"
+    exit 1
+fi
+
+source env/bin/activate
+
 TMPDIR=$(mktemp -d)
 trap "rm -rf $TMPDIR" EXIT
